auth-service/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
email-validator
bcrypt
passlib
pika
starlette
prometheus-client
redis
